# openapi.yml
openapi: 3.0.0
info:
  title: Botapeer API
  description: Botapeer API
  version: 1.0.0
servers:
  - url: "{server}/api"
    variables:
      server:
        default: http://localhost:8081
        enum:
          - http://localhost:8081
          - https://api.botapeer.com
paths:
...
  '/auth/signin':
    post:
      operationId: signin
      tags:
        - Auth
      description: サインイン
      requestBody:
        required: true
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtAuthenticationResponse'
        '404':
          description: 404(User Not Found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/auth/signup':
    post:
      operationId: createUser
      tags:
        - Auth
      description: サインアップ
      requestBody:
        required: true
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/plant_records/{plantRecordId}':
    get:
      operationId: getPlantRecordById
      tags:
        - PlantRecord
      description: 任意の生育記録をplantRecordIdから取得
      parameters:
        - in: path
          name: plantRecordId
          description: 生育記録ID
          required: true
          schema:
            type: string
            example: 1
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantRecordResponse'
        '404':
          description: 404(PlantRecord Not Found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ErrorResponse'
  '/plant_records/users/{userId}':
    get:
      operationId: getPlantRecordByUserId
      tags:
        - PlantRecord
      description: 任意の生育記録をuserIdから取得
      parameters:
        - in: path
          name: userId
          description: ユーザーID
          required: true
          schema:
            type: string
            example: 1
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlantRecordResponse'
        '404':
          description: 404(PlantRecord Not Found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ErrorResponse'
  '/plant_records':
    post:
      operationId: createPlantRecord
      tags:
        - PlantRecord
      description: 生育記録作成
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/CreatePlantRecordRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantRecordResponse'
        '401':
          description: 401(Unauthorized)
          headers:
            http_status:
              description: HTTPステータス
              schema:
                type: integer
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    UserResponse:
      properties:
        id:
          type: integer
        name:
          type: string
          minLength: 1
          maxLength: 15
        email:
          type: string
          format: email
        status:
          type: integer
        description:
          type: string
        profileImage:
          type: string
        coverImage:
          type: string
    PlantRecordResponse:
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          minLength: 1
          maxLength: 40
        alive:
          type: integer
        endDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        place:
            $ref: '#/components/schemas/PlaceResponse'
        posts:
          type: array
          items:
            $ref: '#/components/schemas/PostResponse'
    PlaceResponse:
      properties:
          id:
            type: integer
            format: int64
          name:
            type: string
    PostResponse:
      properties:
          id:
            type: integer
            format: int64
          plantRecordId:
            type: integer
            format: int64
          title:
            type: string
          article:
            type: string
          imageUrl:
            type: string
          status:
            type: integer
          createdAt:
            type: string
            format: date-time
          updatedAt:
            type: string
            format: date-time
          like:
              $ref: '#/components/schemas/LikeResponse'
    LikeResponse:
      properties:
        count:
          type: integer
          format: int64
        isLikeWithRequestUser:
          type: boolean
    SignInRequest:
      properties:
        usernameOrEmail:
          type: string
        password:
          type: string
      required:
        - usernameOrEmail
        - password
    JwtAuthenticationResponse:
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
      required:
        - accessToken
        - tokenType
    CreateUserRequest:
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 15
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 20
      required:
        - name
        - email
        - password
    CreatePlantRecordRequest:
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 40
        placeId:
          type: integer
          format: int64
      required:
        - name
        - email
        - password
...
    ErrorResponse:
      properties:
        errors:
          $ref: '#/components/schemas/Error'
    Error:
        type: array
        items:
          properties:
            code:
              type: integer
            message:
              type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT